+++ spec/unit/target/aggregate_target_spec.rb
@@ -224,12 +224,6 @@ module Pod
             @target.requires_host_target?.should == true
           end
 
-          it 'does not require a host target for messages extension targets embedded in messages applications' do
-            @target.add_host_target_product_type(:messages_application)
-            @target.user_targets.first.stubs(:symbol_type).returns(:messages_extension)
-            @target.requires_host_target?.should == false
-          end
-
           it 'does not require a host target for watch 2 extension targets' do
             @target.user_targets.first.stubs(:symbol_type).returns(:watch2_extension)
             @target.requires_host_target?.should == false